Intermediate commit aith support added for AVR (GPIO,ADC). Does not support other targets.

+++ src/c/scilab-arduino/interfaces/int_cmd_digital_out.h
@@ -25,7 +25,7 @@ extern "C" {
 
 //Required when input from one pin is directed to other output pin.
 #define d0d0u80cmd_digital_outu80(in1,in2,in3) u8cmd_digital_outs((uint8)in1,\
-						(uint8)in2, in3)	
+						(uint8)in2, (uint8)in3)	
 
 
 #ifdef  __cplusplus
